+++ instrumentation/opentelemetry-instrumentation-httpx/src/opentelemetry/instrumentation/httpx/__init__.py
@@ -233,6 +233,38 @@ def _prepare_headers(headers: typing.Optional[Headers]) -> httpx.Headers:
     return httpx.Headers(headers)
 
 
+def _extract_parameters(args, kwargs):
+    if isinstance(args[0], httpx.Request):
+        # In httpx >= 0.20.0, handle_request receives a Request object
+        request: httpx.Request = args[0]
+        method = request.method.encode()
+        url = request.url
+        headers = request.headers
+        stream = request.stream
+        extensions = request.extensions
+    else:
+        # In httpx < 0.20.0, handle_request receives the parameters separately
+        method = args[0]
+        url = args[1]
+        headers = kwargs.get("headers", args[2] if len(args) > 2 else None)
+        stream = kwargs.get("stream", args[3] if len(args) > 3 else None)
+        extensions = kwargs.get(
+            "extensions", args[4] if len(args) > 4 else None
+        )
+
+    return method, url, headers, stream, extensions
+
+
+def _inject_propagation_headers(headers, args, kwargs):
+    _headers = _prepare_headers(headers)
+    inject(_headers)
+    if isinstance(args[0], httpx.Request):
+        request: httpx.Request = args[0]
+        request.headers = _headers
+    else:
+        kwargs["headers"] = _headers.raw
+
+
 class SyncOpenTelemetryTransport(httpx.BaseTransport):
     """Sync transport class that will trace all requests made with a client.
 
@@ -263,60 +295,53 @@ class SyncOpenTelemetryTransport(httpx.BaseTransport):
 
     def handle_request(
         self,
-        method: bytes,
-        url: URL,
-        headers: typing.Optional[Headers] = None,
-        stream: typing.Optional[httpx.SyncByteStream] = None,
-        extensions: typing.Optional[dict] = None,
-    ) -> typing.Tuple[int, "Headers", httpx.SyncByteStream, dict]:
+        *args,
+        **kwargs,
+    ) -> typing.Union[
+        typing.Tuple[int, "Headers", httpx.SyncByteStream, dict],
+        httpx.Response,
+    ]:
         """Add request info to span."""
         if context.get_value("suppress_instrumentation"):
-            return self._transport.handle_request(
-                method,
-                url,
-                headers=headers,
-                stream=stream,
-                extensions=extensions,
-            )
+            return self._transport.handle_request(*args, **kwargs)
 
+        method, url, headers, stream, extensions = _extract_parameters(
+            args, kwargs
+        )
         span_attributes = _prepare_attributes(method, url)
-        _headers = _prepare_headers(headers)
+
+        request_info = RequestInfo(method, url, headers, stream, extensions)
         span_name = _get_default_span_name(
             span_attributes[SpanAttributes.HTTP_METHOD]
         )
-        request = RequestInfo(method, url, headers, stream, extensions)
 
         with self._tracer.start_as_current_span(
             span_name, kind=SpanKind.CLIENT, attributes=span_attributes
         ) as span:
             if self._request_hook is not None:
-                self._request_hook(span, request)
+                self._request_hook(span, request_info)
 
-            inject(_headers)
-
-            (
-                status_code,
-                headers,
-                stream,
-                extensions,
-            ) = self._transport.handle_request(
-                method,
-                url,
-                headers=_headers.raw,
-                stream=stream,
-                extensions=extensions,
-            )
+            _inject_propagation_headers(headers, args, kwargs)
+            response = self._transport.handle_request(*args, **kwargs)
+            if isinstance(response, httpx.Response):
+                response: httpx.Response = response
+                status_code = response.status_code
+                headers = response.headers
+                stream = response.stream
+                extensions = response.extensions
+            else:
+                status_code, headers, stream, extensions = response
 
             _apply_status_code(span, status_code)
 
             if self._response_hook is not None:
                 self._response_hook(
                     span,
-                    request,
+                    request_info,
                     ResponseInfo(status_code, headers, stream, extensions),
                 )
 
-        return status_code, headers, stream, extensions
+        return response
 
 
 class AsyncOpenTelemetryTransport(httpx.AsyncBaseTransport):
@@ -348,61 +373,55 @@ class AsyncOpenTelemetryTransport(httpx.AsyncBaseTransport):
         self._response_hook = response_hook
 
     async def handle_async_request(
-        self,
-        method: bytes,
-        url: URL,
-        headers: typing.Optional[Headers] = None,
-        stream: typing.Optional[httpx.AsyncByteStream] = None,
-        extensions: typing.Optional[dict] = None,
-    ) -> typing.Tuple[int, "Headers", httpx.AsyncByteStream, dict]:
+        self, *args, **kwargs
+    ) -> typing.Union[
+        typing.Tuple[int, "Headers", httpx.AsyncByteStream, dict],
+        httpx.Response,
+    ]:
         """Add request info to span."""
         if context.get_value("suppress_instrumentation"):
-            return await self._transport.handle_async_request(
-                method,
-                url,
-                headers=headers,
-                stream=stream,
-                extensions=extensions,
-            )
+            return await self._transport.handle_async_request(*args, **kwargs)
 
+        method, url, headers, stream, extensions = _extract_parameters(
+            args, kwargs
+        )
         span_attributes = _prepare_attributes(method, url)
-        _headers = _prepare_headers(headers)
+
         span_name = _get_default_span_name(
             span_attributes[SpanAttributes.HTTP_METHOD]
         )
-        request = RequestInfo(method, url, headers, stream, extensions)
+        request_info = RequestInfo(method, url, headers, stream, extensions)
 
         with self._tracer.start_as_current_span(
             span_name, kind=SpanKind.CLIENT, attributes=span_attributes
         ) as span:
             if self._request_hook is not None:
-                await self._request_hook(span, request)
+                await self._request_hook(span, request_info)
 
-            inject(_headers)
+            _inject_propagation_headers(headers, args, kwargs)
 
-            (
-                status_code,
-                headers,
-                stream,
-                extensions,
-            ) = await self._transport.handle_async_request(
-                method,
-                url,
-                headers=_headers.raw,
-                stream=stream,
-                extensions=extensions,
+            response = await self._transport.handle_async_request(
+                *args, **kwargs
             )
+            if isinstance(response, httpx.Response):
+                response: httpx.Response = response
+                status_code = response.status_code
+                headers = response.headers
+                stream = response.stream
+                extensions = response.extensions
+            else:
+                status_code, headers, stream, extensions = response
 
             _apply_status_code(span, status_code)
 
             if self._response_hook is not None:
                 await self._response_hook(
                     span,
-                    request,
+                    request_info,
                     ResponseInfo(status_code, headers, stream, extensions),
                 )
 
-        return status_code, headers, stream, extensions
+        return response
 
 
 class _InstrumentedClient(httpx.Client):