+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/bootstrap_gen.py
@@ -69,7 +69,7 @@ libraries = {
         "instrumentation": "opentelemetry-instrumentation-grpc==0.28b1",
     },
     "httpx": {
-        "library": "httpx >= 0.18.0, < 0.19.0",
+        "library": "httpx >= 0.18.0",
         "instrumentation": "opentelemetry-instrumentation-httpx==0.28b1",
     },
     "jinja2": {